+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2025-04-07
+- update GitHub actions
+- update readme.md
+- add **int lastError()** to prepare error handling
+- add **uint32_t lastRead()**
+- 
+
+
 ## [0.1.0] - 2023-12-24
 - initial release